Adjust ppc64-diag yaml and spec to be flexible to version changes
Right now the ppc64-diag yaml was tied to an source tarball with 2.7.1 version
on its name. Removing this also requires to change the name of the expected
Source0 file in its spec.

Also adjustment is necessary in /ppc64-diag-2.4.3-scriptlocation.patch to make
it apply in the current repository state