Infer the essential sources tar.gz file name (archive attrib) instead of mentioning it explicitly in package yaml file
When creating the essential sources tar.gz from a git repository, there is no reason to use a
file name different from the package name, so this attribute is useless in this case.

When downloading the essential sources tar.gz instead of creating it from a git repository,
archive attribute may be useful. This currently happens in gcc, kubernetes and libguestfs.
In kubernetes and libguestfs, a tar.gz is downloaded using wget, so archive attribute was
dropped by passing an output file name in standard format used by OpenPower Host OS build
script to the wget command. In gcc, a tar.bz2 is created instead of a tar.gz file, so the
archive attribute must be kept. It should be evaluated if using tar.bz2 instead of
tar.gz in gcc is essential and, if it is not, drop it for good in a future patch.