Provide qemu-kvm-rhev in qemu package
CentOS 7.3 now has qemu-kvm-ev and libguestfs in the extras
repo. libguestfs asks for qemu-kvm-rhev, which is not provided by our
qemu package. Up until now we've been altering our libguestfs spec
file to skip that requirement for ppc64le. Since we intent to use the
libguestfs package from the distro, we decided to make our qemu
package provide qemu-kvm-rhel to satisfy the requirement.