Change kubernetes URL source to a git source
This makes it consistent with other packages.
Remove the URL from its main source, since its created from the git
repository and not downloaded directly anymore.

The project part of the URL had to be changed from kubernetes to
openshift because the specific commit built is not available from the
main kubernetes repository branches.