Change kubernetes secondary sources to git repositories
This is more efficient than re-downloading the archive files for every
build.
Remove the URLs from the spec file, since those files are not downloaded
directly anymore.
The archive names needed to change to the name of the extracted
directory that the spec file expects. This avoids further modifications
to the spec file.