Rename noVNC directory and package name in yaml to novnc (package name in rpm spec)
Directory name and package name in yaml should always be identical
to the package name in the rpm specification file.

Although the git repository URL is case-independent, the git repository name
is kept as noVNC, as the git repo name is not related to the package name or
tar.gz file name.