Add -q to %setup macro in rpm specification files
Some rpm packages used the -q option to avoid useless build output
when unpacking the sources, others not. This patch makes it consistent
in all rpm packages.

Also, it is a rpmlint check (exactly because of the reason given above)
and the less rpmlint checks we need to filter, better.