Provide qemu-*-ev instead of requiring them
This avoids conflicts for upstream packages which require those names
(libguestfs).

There is currently one libguestfs package in CentOS updates repository
and one in extras repository. When we try to install it, yum is
selecting the one from updates repository. Although they appear to be
identical, for some unknown reason, yum tools list different
requirements for each one of them. In particular, the one from updates
lists qemu-img-ev as a requirement, while the one from extras lists
qemu-img instead. Since our qemu spec file was obsoleting qemu-img-ev
(and it is the package providing some of the other requirements for
libguestfs), yum was unable to fulfill that requirement.

By providing qemu-img-ev instead of obsoleting it, we fulfill the
reuquirement for libguestfs and still avoid installing the old upstream
qemu package. The same is done for other qemu subpackages, since a
similar issue could eventually arise from them.