Fix open-power-host-os-virt-management obsoleted version
Since yum uses a lexical comparison between package versions, the
version we were previously obsoleting was lesser than the last release
of the open-power-host-os-virt-management, and yum was uanble to resolve
conflicts when trying to install the new packages, which do not contain
open-power-host-os-virt-management.

By incrementing the obsoleted version, we make sure any such package
will be obsoleted, regardless of the milestone marked in the release
field and dist in which it was built.