Remove skopeo package
Since docker is no longer built by us, we don't need to keep building
this package, which was added because it is a dependency of docker.
CentOS 7.4 now also contains skopeo-0.1.20-2.1.gite802625.el7.ppc64le,
which is the same version we had, but a lesser release. It is recommended
to uninstall skopeo from previous Host OS builds.