qemu: force `--with kvmonly` as default behavior
Since CentOS 7.5, sgabios-bin is not provided in the os/ repository. This arose
an issue in our qemu.spec, which mistakenly builds qemu-system-x86 on ppc64
architecture.

This patch assumes `--with kvmonly` as the default behavior, which won't build
qemu-system-x86 anymore; and also fixes some non-x86 files packaging.

Even though we're forcing `--with kvmonly`, we still provide qemu-system-ppc
package for compatibility reasons.