kernel: revert changes incompatible with current build system
The commit "a670b0b kbuild: Use ls(1) instead of stat(1) to obtain
file size" includes a change in arch/powerpc/boot/wrapper that expects
a kernel source tree to be present:

-strip_size=$(stat -c %s $vmz.$$)
+strip_size=$(${CONFIG_SHELL} "${srctree}/scripts/file-size.sh" "$vmz.$$")

However, during our ISO build process, lorax calls this script to
prepare the boot images without having a source tree setup. This
results in the following error:

/work/ppc64le/installroot/usr/sbin/wrapper: line 358:
/scripts/file-size.sh: No such file or directory

The upstream solution for this bug was to alter lorax' behavior to
avoid using the wrapper:

https://bugzilla.redhat.com/show_bug.cgi?id=1566225

Since we currently use an older version of lorax (19.6.104, from
CentOS 7.5) and the kernel change that introduced this issue is simple
enough, we'll just carry this patch until the base distro updates their
lorax version.